Update to rspec3 conventions
