+++ spec/controllers/application_controller_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe ApplicationController, type: :controller do
    describe "GET 'index'" do
