+++ spec/jobs/update_dataset_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe UpdateDataset do
