+++ spec/models/dataset_file_schema_spec.rb
@@ -1,6 +1,6 @@
 # == Schema Information
 
-require 'spec_helper'
+require 'rails_helper'
 
 describe DatasetFileSchema do
