+++ spec/models/dataset_spec.rb
@@ -24,7 +24,7 @@
 #  restricted      :boolean          default(FALSE)
 #
 
-require 'spec_helper'
+require 'rails_helper'
 
 describe Dataset do
