+++ spec/models/git_data_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe GitData, :vcr do
