+++ spec/models/user_spec.rb
@@ -15,7 +15,7 @@
 #  twitter_handle  :string
 #
 
-require 'spec_helper'
+require 'rails_helper'
 
 describe User do
