+++ spec/rails_helper.rb
@@ -1,0 +1,150 @@
+# This file is copied to spec/ when you run 'rails generate rspec:install'
+ENV['RAILS_ENV'] ||= 'test'
+require File.expand_path('../../config/environment', __FILE__)
+# Prevent database truncation if the environment is production
+abort("The Rails environment is running in production mode!") if Rails.env.production?
+
+require 'coveralls'
+Coveralls.wear!('rails')
+
+require 'spec_helper'
+require 'rspec/rails'
+
+require 'git_data'
+require 'database_cleaner'
+require 'factory_girl'
+require 'omniauth'
+require 'support/vcr_helper'
+require 'support/fake_data'
+require 'webmock/rspec'
+require 'sidekiq/testing'
+
+
+# Add additional requires below this line. Rails is not loaded until this point!
+
+# Requires supporting ruby files with custom matchers and macros, etc, in
+# spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
+# run as spec files by default. This means that files in spec/support that end
+# in _spec.rb will both be required and run as specs, causing the specs to be
+# run twice. It is recommended that you do not name files matching this glob to
+# end with _spec.rb. You can configure this pattern with the --pattern
+# option on the command line or in ~/.rspec, .rspec or `.rspec-local`.
+#
+# The following line is provided for convenience purposes. It has the downside
+# of increasing the boot-up time by auto-requiring all files in the support
+# directory. Alternatively, in the individual `*_spec.rb` files, manually
+# require only the support files necessary.
+#
+# Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
+
+# Checks for pending migration and applies them before tests are run.
+# If you are not using ActiveRecord, you can remove this line.
+ActiveRecord::Migration.maintain_test_schema!
+
+RSpec.configure do |config|
+  # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
+  config.fixture_path = "#{::Rails.root}/spec/fixtures"
+
+  # If you're not using ActiveRecord, or you'd prefer not to run each of your
+  # examples within a transaction, remove the following line or assign false
+  # instead of true.
+  config.use_transactional_fixtures = true
+
+  config.include FactoryGirl::Syntax::Methods
+  # RSpec Rails can automatically mix in different behaviours to your tests
+  # based on their file location, for example enabling you to call `get` and
+  # `post` in specs under `spec/controllers`.
+  #
+  # You can disable this behaviour by removing the line below, and instead
+  # explicitly tag your specs with their type, e.g.:
+  #
+  #     RSpec.describe UsersController, :type => :controller do
+  #       # ...
+  #     end
+  #
+  # The different available types are documented in the features, such as in
+  # https://relishapp.com/rspec/rspec-rails/docs
+  config.infer_spec_type_from_file_location!
+
+  # Filter lines from Rails gems in backtraces.
+  config.filter_rails_from_backtrace!
+  # arbitrary gems may also be filtered via:
+  # config.filter_gems_from_backtrace("gem name")
+
+  config.before(:each) do |example|
+    # Stub out repository checking for all tests apart from GitData
+    allow_any_instance_of(Octokit::Client).to receive(:repository?) { false } unless example.metadata[:described_class] == GitData
+  end
+
+  # This overrides always true in the spec_helper file
+  config.around(:each, type: :helper) do |ex|
+    config.mock_with :rspec do |mocks|
+      mocks.verify_partial_doubles = false
+      ex.run
+      mocks.verify_partial_doubles = true
+    end
+  end
+end
+
+def sign_in(user)
+  allow_any_instance_of(ApplicationController).to receive(:session) { {user_id: user.id} }
+end
+
+def sign_out
+  allow_any_instance_of(ApplicationController).to receive(:session) { {} }
+end
+
+def set_api_key(user)
+  sign_out # Make sure we haven't got any sessions hanging around
+  request.env['HTTP_AUTHORIZATION'] = ActionController::HttpAuthentication::Token.encode_credentials(user.api_key)
+end
+
+def get_json_from_url(url)
+  JSON.generate(JSON.load(open(url).read.force_encoding("UTF-8")))
+end
+
+def url_with_stubbed_get_for(path)
+  url = "https://example.org/uploads/#{SecureRandom.uuid}/somefile.csv"
+  stub_request(:get, url).to_return(body: File.read(path))
+  url
+end
+
+def url_for_schema_with_stubbed_get_for(path)
+  url = "https://example.org/uploads/#{SecureRandom.uuid}/schema.json"
+  stub_request(:get, url).to_return(body: File.read(path))
+  url
+end
+
+def mock_pusher(channel_id)
+  mock_client = double(Pusher::Channel)
+  expect(Pusher).to receive(:[]).with(channel_id) { mock_client }
+  mock_client
+end
+
+def skip_dataset_callbacks!
+  skip_callback_if_exists(Dataset, :create, :after, :create_repo_and_populate)
+  skip_callback_if_exists(Dataset, :create, :after, :set_owner_avatar)
+  skip_callback_if_exists(Dataset, :create, :after, :publish_public_views)
+  skip_callback_if_exists(Dataset, :create, :after, :send_success_email)
+end
+
+def set_dataset_callbacks!
+  Dataset.set_callback(:create, :after, :create_repo_and_populate)
+  Dataset.set_callback(:create, :after, :set_owner_avatar)
+  Dataset.set_callback(:create, :after, :publish_public_views)
+  Dataset.set_callback(:create, :after, :send_success_email)
+end
+
+
+def skip_callback_if_exists(thing, name, kind, filter)
+  if name == :create && any_callbacks?(thing._create_callbacks, name, kind, filter)
+    thing.skip_callback(name, kind, filter)
+  end
+  if name == :update && any_callbacks?(thing._update_callbacks, name, kind, filter)
+    thing.skip_callback(name, kind, filter)
+  end
+end
+
+def any_callbacks?(callbacks, name, kind, filter)
+  callbacks.select { |cb| cb.name == name && cb.kind == kind && cb.filter == filter }.any?
+end
