+++ spec/services/dataset_file_schema_service_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe DatasetFileSchemaService do
