+++ spec/spec_helper.rb
@@ -1,121 +1,113 @@
-ENV['RAILS_ENV'] ||= 'test'
-
-require 'coveralls'
-Coveralls.wear!('rails')
-
-require File.expand_path('../../config/environment', __FILE__)
-abort("The Rails environment is running in production mode!") if Rails.env.production?
-
-require 'git_data'
-
-require 'rspec/rails'
-require 'database_cleaner'
-require 'factory_girl'
-require 'omniauth'
-require 'support/vcr_helper'
-require 'support/fake_data'
-require 'webmock/rspec'
-require 'sidekiq/testing'
-
-DatabaseCleaner.strategy = :truncation
-OmniAuth.config.test_mode = true
-
-ActiveRecord::Migration.maintain_test_schema!
-
+# This file was generated by the `rails generate rspec:install` command. Conventionally, all
+# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
+# The generated `.rspec` file contains `--require spec_helper` which will cause
+# this file to always be loaded, without a need to explicitly require it in any
+# files.
+#
+# Given that it is always loaded, you are encouraged to keep this file as
+# light-weight as possible. Requiring heavyweight dependencies from this file
+# will add to the boot time of your test suite on EVERY test run, even for an
+# individual file that may not need all of that loaded. Instead, consider making
+# a separate helper file that requires the additional dependencies and performs
+# the additional setup, and require it from the spec files that actually need
+# it.
+#
+# The `.rspec` file also contains a few flags that are not defaults but that
+# users commonly want.
+#
+# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
-  config.fixture_path = "#{::Rails.root}/spec/fixtures"
-  config.use_transactional_fixtures = true
-  config.include FactoryGirl::Syntax::Methods
-
+  # rspec-expectations config goes here. You can use an alternate
+  # assertion/expectation library such as wrong or the stdlib/minitest
+  # assertions if you prefer.
   config.expect_with :rspec do |expectations|
+    # This option will default to `true` in RSpec 4. It makes the `description`
+    # and `failure_message` of custom matchers include text for helper methods
+    # defined using `chain`, e.g.:
+    #     be_bigger_than(2).and_smaller_than(4).description
+    #     # => "be bigger than 2 and smaller than 4"
+    # ...rather than:
+    #     # => "be bigger than 2"
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
+  # rspec-mocks config goes here. You can use an alternate test double
+  # library (such as bogus or mocha) by changing the `mock_with` option here.
   config.mock_with :rspec do |mocks|
+    # Prevents you from mocking or stubbing a method that does not exist on
+    # a real object. This is generally recommended, and will default to
+    # `true` in RSpec 4.
     mocks.verify_partial_doubles = true
   end
 
-  config.order = :random
+  # This option will default to `:apply_to_host_groups` in RSpec 4 (and will
+  # have no way to turn it off -- the option exists only for backwards
+  # compatibility in RSpec 3). It causes shared context metadata to be
+  # inherited by the metadata hash of host groups and examples, rather than
+  # triggering implicit auto-inclusion in groups with matching metadata.
+  config.shared_context_metadata_behavior = :apply_to_host_groups
 
-  config.before(:each) do |example|
-    # Stub out repository checking for all tests apart from GitData
-    allow_any_instance_of(Octokit::Client).to receive(:repository?) { false } unless example.metadata[:described_class] == GitData
+  config.expect_with :rspec do |expectations|
+    expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
   config.after(:each) do
     DatabaseCleaner.clean
   end
 
-  config.around(:each, type: :helper) do |ex|
-    config.mock_with :rspec do |mocks|
-      mocks.verify_partial_doubles = false
-      ex.run
-      mocks.verify_partial_doubles = true
-    end
+
+
+  # Run specs in random order to surface order dependencies. If you find an
+  # order dependency and want to debug it, you can fix the order by providing
+  # the seed, which is printed after each run.
+  #     --seed 1234
+  config.order = :random
+
+
+# The settings below are suggested to provide a good initial experience
+# with RSpec, but feel free to customize to your heart's content.
+=begin
+  # This allows you to limit a spec run to individual examples or groups
+  # you care about by tagging them with `:focus` metadata. When nothing
+  # is tagged with `:focus`, all examples get run. RSpec also provides
+  # aliases for `it`, `describe`, and `context` that include `:focus`
+  # metadata: `fit`, `fdescribe` and `fcontext`, respectively.
+  config.filter_run_when_matching :focus
+
+  # Allows RSpec to persist some state between runs in order to support
+  # the `--only-failures` and `--next-failure` CLI options. We recommend
+  # you configure your source control system to ignore this file.
+  config.example_status_persistence_file_path = "spec/examples.txt"
+
+  # Limits the available syntax to the non-monkey patched syntax that is
+  # recommended. For more details, see:
+  #   - http://rspec.info/blog/2012/06/rspecs-new-expectation-syntax/
+  #   - http://www.teaisaweso.me/blog/2013/05/27/rspecs-new-message-expectation-syntax/
+  #   - http://rspec.info/blog/2014/05/notable-changes-in-rspec-3/#zero-monkey-patching-mode
+  config.disable_monkey_patching!
+
+  # Many RSpec users commonly either run the entire suite or an individual
+  # file, and it's useful to allow more verbose output when running an
+  # individual spec file.
+  if config.files_to_run.one?
+    # Use the documentation formatter for detailed output,
+    # unless a formatter has already been configured
+    # (e.g. via a command-line flag).
+    config.default_formatter = 'doc'
   end
 
-  config.infer_spec_type_from_file_location!
-end
+  # Print the 10 slowest examples and example groups at the
+  # end of the spec run, to help surface which specs are running
+  # particularly slow.
+  config.profile_examples = 10
 
-def sign_in(user)
-  allow_any_instance_of(ApplicationController).to receive(:session) { {user_id: user.id} }
-end
-
-def sign_out
-  allow_any_instance_of(ApplicationController).to receive(:session) { {} }
-end
-
-def set_api_key(user)
-  sign_out # Make sure we haven't got any sessions hanging around
-  request.env['HTTP_AUTHORIZATION'] = ActionController::HttpAuthentication::Token.encode_credentials(user.api_key)
-end
-
-def get_json_from_url(url)
-  JSON.generate(JSON.load(open(url).read.force_encoding("UTF-8")))
-end
-
-def url_with_stubbed_get_for(path)
-  url = "https://example.org/uploads/#{SecureRandom.uuid}/somefile.csv"
-  stub_request(:get, url).to_return(body: File.read(path))
-  url
-end
-
-def url_for_schema_with_stubbed_get_for(path)
-  url = "https://example.org/uploads/#{SecureRandom.uuid}/schema.json"
-  stub_request(:get, url).to_return(body: File.read(path))
-  url
-end
-
-def mock_pusher(channel_id)
-  mock_client = double(Pusher::Channel)
-  expect(Pusher).to receive(:[]).with(channel_id) { mock_client }
-  mock_client
-end
-
-def skip_dataset_callbacks!
-  skip_callback_if_exists(Dataset, :create, :after, :create_repo_and_populate)
-  skip_callback_if_exists(Dataset, :create, :after, :set_owner_avatar)
-  skip_callback_if_exists(Dataset, :create, :after, :publish_public_views)
-  skip_callback_if_exists(Dataset, :create, :after, :send_success_email)
-end
-
-def set_dataset_callbacks!
-  Dataset.set_callback(:create, :after, :create_repo_and_populate)
-  Dataset.set_callback(:create, :after, :set_owner_avatar)
-  Dataset.set_callback(:create, :after, :publish_public_views)
-  Dataset.set_callback(:create, :after, :send_success_email)
+  # Seed global randomization in this process using the `--seed` CLI option.
+  # Setting this allows you to use `--seed` to deterministically reproduce
+  # test failures related to randomization by passing the same `--seed` value
+  # as the one that triggered the failure.
+  Kernel.srand config.seed
+=end
 end
 
 
-def skip_callback_if_exists(thing, name, kind, filter)
-  if name == :create && any_callbacks?(thing._create_callbacks, name, kind, filter)
-    thing.skip_callback(name, kind, filter)
-  end
-  if name == :update && any_callbacks?(thing._update_callbacks, name, kind, filter)
-    thing.skip_callback(name, kind, filter)
-  end
-end
 
-def any_callbacks?(callbacks, name, kind, filter)
-  callbacks.select { |cb| cb.name == name && cb.kind == kind && cb.filter == filter }.any?
-end
