+++ spec/views/_dataset_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe 'datasets/_dataset.html.erb' do
