+++ spec/views/_datasets_spec.rb
@@ -1,4 +1,4 @@
-require 'spec_helper'
+require 'rails_helper'
 
 describe 'datasets/_datasets.html.erb' do
